Added exception and its handling

+++ HR.LeaveManagement.Application/Features/LeaveTypes/Handlers/Commands/DeleteLeaveTypeCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using HR.LeaveManagement.Application.Exceptions;
 using HR.LeaveManagement.Application.Features.LeaveTypes.Request.Commands;
 using HR.LeaveManagement.Application.Persistence.Contracts;
 using MediatR;
@@ -16,9 +17,11 @@ namespace HR.LeaveManagement.Application.Features.LeaveTypes.Handlers.Commands
         }
 
         public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
-        {
+        {   
             var leaveType = await _leaveTypeRepository.Get(request.Id);
 
+            if(leaveType == null)
+                throw new NotFoundException(nameof(leaveType), request.Id);
 
             await _leaveTypeRepository.Delete(leaveType);
